backend/requirements.txt
Flask
requests
python-dotenv
numpy
SQLAlchemy
Flask-SQLAlchemy
Faker
airportsdata
psycopg2-binary
flask-cors
pycountry
pytest 
flask-testing
locust
Flask-JWT-Extended
bcrypt
coverage 